Describe Video Function/requirements.txt
functions-framework
google-cloud-storage
google-cloud-functions
google-cloud-aiplatform
flask